Fix test runner script
Before it would exit with code 0, even if unit tests failed. To properly
report failing unit tests we simply add the exit code of the end2end
test runner to the exit code of the unit test runner. It will return 0
only if both returned 0 since both return 1 in case of failure.