Add postinstall-build to build after an install
This is used for people who want to install js-module-walker directly
from github. Right now npm won't run the prepublish hook so nothing is
transpiled and js-module-walker cannot be run.

This package runs as postinstall and checks whether the build folder
exists. If it does not exist it will install all dev dependencies and
run the build script.